A page in the game sets every key, mouse button and pad button
The Controls page has a ninth row, Buttons, and A on it opens the BUTTONS
page: a Device row (Keyboard, Mouse, Controller; Left and Right pick it,
the Controller first when a pad is attached), a row for each of the
game's eighteen inputs showing what presses it on that device, and Reset
All. A on an input row makes the port listen for the next press of that
device and binds it in place of the device's old one; Z clears the row on
that device, refused when nothing else would press the input; Esc or six
seconds of nothing leaves the row as it was; a key the port answers to
itself (the function keys, P, the brackets, Home, End, Esc) and the pad's
Back and Guide are refused with a line saying so. Every change is in
force at once and reaches the settings file by the usual debounced write,
into the same keys table a text editor changes. Asked for in Discussions
#3, and the first of the things the other N64 recompilations have that
this port did not (the survey of Sep 13 2026).

The page runs on the game's thread, like the other pages, and talks to the
host through a new bank of the mailbox (+0xA0: a request word, an answer
word, a generation word, the device shown, open, pad attached). The row
values cannot be staged ahead of time -- any key on the keyboard may be
named -- so the host composes them live for the device shown, from the
binding table, into two banks of eighteen ids whose pixels sit apart from
the staged strings (0x80D00000); the generation word's low bit names the
bank the page swaps its strips to, so a bank is never drawn while it is
rewritten. While a capture listens the input layer hands the game no
input, and none until every key and button is let go after, so the press
cannot land as the button it was or as the one it now is; the window's
thread hands every key, mouse button, wheel tick, pad button and trigger
edge to the capture first, and skips the hotkeys and Esc's Start while
one is on. The body face gained the capitals I K Q U X Y, a q and the
punctuation a key can be named with, drawn in its style, and the header
face a B for the heading; a glyph no face has draws as a dim block in a
live value and never withholds the directory.

Two latent faults found on the way: the menu directory seated 127 ids and
1.0.5 staged 129, the last two entries landing in the black tile's pixels
that nothing draws (the pixels start at 0x80C01800 now, 255 ids); and the
strip pool's two counters shared +0x70/+0x74 with the Exit Game item's
pointers, harmless only because a pointer outranks any count. The
SNAP_PASS_TRACE line printed a 64-bit workload id through %u.

Verified by replay on Windows (scratchpad make_bind_replay.py and
bind_probe.py, a scratch data directory, SNAP_BIND_TEST feeding the
presses): Space bound to A, B's keyboard key cleared, Stick Up's clear
refused, the device turned to Controller, Pad Y bound to A in place of
Pad A, back out to the title; the file then held a = Mouse Left, Space,
Pad Y and b = Mouse Middle, Pad B, Pad X, with stick_up untouched. The
event decode of a real key or pad press, and the look of the page, are
not covered by that run.

+++ lib/rt64/src/hle/rt64_workload_queue.cpp
@@ -827,8 +827,8 @@ namespace RT64 {
                                 traceThis = tiny && (tracedWorkloads < 40);
                                 if (traceThis) {
                                     tracedWorkloads++;
-                                    fprintf(stdout, "[SNAP-PASS] workload %u: %u pairs, presented pair %d, msaa %d, interpolated %d\n",
-                                        workload.workloadId, fbPairCount, overrideTargetFbPairIndex, usingMSAA ? 1 : 0, interpolationSubFrame ? 1 : 0);
+                                    fprintf(stdout, "[SNAP-PASS] workload %llu: %u pairs, presented pair %d, msaa %d, interpolated %d\n",
+                                        (unsigned long long) workload.workloadId, fbPairCount, overrideTargetFbPairIndex, usingMSAA ? 1 : 0, interpolationSubFrame ? 1 : 0);
                                 }
                             }
                             if (traceThis) {
